# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CPhOS 物理竞赛题生成 API",
    "description": "\nCPhOS 物理竞赛题全自动生成系统的后端 API。\n\n## 鉴权\n所有 `/api/**` 端点需在 `Authorization: Bearer <token>` 头中携带 token。\ntoken 由管理员通过 `/api/admin/tokens` 签发，明文仅创建时返回一次。\n\n## 异步工作流\n1. `POST /api/tasks` 提交任务，立即获得 `task_id`；\n2. 轮询 `GET /api/tasks/{task_id}` 直到 `status` 变为 `done` / `error` / `aborted`；\n3. 通过 `GET /api/tasks/{task_id}/artifacts` 与下载端点获取产物。\n",
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "返回当前身份与角色",
        "description": "返回当前 token 对应的用户 ID、角色与备注名，可用于登录后路由与 token 校验。",
        "operationId": "get_me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "提交生成任务",
        "description": "提交一个异步生成任务，立即返回 ``task_id``；通过轮询获取进度与结果。",
        "operationId": "create_task_api_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreated"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "列出当前用户的任务历史",
        "description": "分页列出调用者自己的历史任务，支持状态 / 模式 / 主题过滤与排序。",
        "operationId": "list_my_tasks_api_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤，可逗号分隔（如 running,error）。",
              "title": "Status"
            },
            "description": "按状态过滤，可逗号分隔（如 running,error）。"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按命题模式过滤。",
              "title": "Mode"
            },
            "description": "按命题模式过滤。"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按 topic 模糊搜索。",
              "title": "Q"
            },
            "description": "按 topic 模糊搜索。"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "按创建时间排序：ASC / DESC。",
              "default": "DESC",
              "title": "Order"
            },
            "description": "按创建时间排序：ASC / DESC。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TaskListItem_"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/upload": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "上传源材料并提交改编任务",
        "description": "上传源材料文件并提交改编任务。",
        "operationId": "create_task_upload_api_tasks_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_task_upload_api_tasks_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreated"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/tasks/{task_id}/progress": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询任务的节点级进度",
        "description": "返回任务的阶段事件时间线，含每个已完成阶段的结构化产出快照。\n\n可在轮询任务期间调用，用于展示当前所处节点与各阶段的格式化结果\n（非模型原始输出）。",
        "operationId": "get_task_progress_api_tasks__task_id__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskProgress"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}/artifacts/archive": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "打包下载全部产物（zip）",
        "description": "把任务的全部产物打包为单个 zip 流式返回，便于「一键下载全部」。\n\n压缩包内文件名沿用产物的磁盘名（含 ``{task_id}_assets/`` 子目录结构）。",
        "operationId": "download_artifacts_archive_api_tasks__task_id__artifacts_archive_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "全部产物的 zip 压缩包，附 Content-Disposition 文件名。",
            "content": {
              "application/zip": {}
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "当前 token 所属用户 ID。"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "title": "Role",
            "description": "当前 token 角色。"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "用户备注名。",
            "default": ""
          },
          "token_id": {
            "type": "string",
            "title": "Token Id",
            "description": "当前会话所用 token 的 ID。"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role",
          "token_id"
        ],
        "title": "MeResponse",
        "description": "当前调用者的身份与角色（任意有效 token 均可访问）。"
      },
...
      "TaskCreated": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "任务标识，用于后续轮询与下载。"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务初始状态。",
            "examples": [
              "queued"
            ]
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskCreated",
        "description": "任务提交后的即时响应。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "不透明 token，形如 `Bearer <token>`",
        "scheme": "bearer"
      }
    }
  }
}